# docs/api/project-settings.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Project settings — review S03/S10",
    "version": "0.10.1"
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/bug-rule-versions": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "listProjectRulesets",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ruleset"
                  }
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProjectRuleset",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ruleset"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleset"
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/bug-rule-versions/{rulesetId}/versions": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        },
        {
          "name": "rulesetId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "get": {
        "operationId": "listProjectRuleVersions",
        "description": "",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleVersion"
                  }
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProjectRuleVersion",
        "description": "",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleVersion"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleVersion"
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/bug-rule-versions/{rulesetId}/versions/{versionId}/publish": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        },
        {
          "name": "rulesetId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        },
        {
          "name": "versionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      ],
      "post": {
        "operationId": "publishProjectRuleVersion",
        "description": "0 là chưa có version active. Chỉ quy tắc nội bộ DEMO; không thay quy tắc khách hàng.",
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ruleset"
                }
              }
            }
          },
          "default": {
            "description": "401 phiên; 403 quyền; 404 phạm vi; 422 input; 409 version/trùng/PM cuối",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "maxLength": 80
                    },
                    "message": {
                      "type": "string",
                      "maxLength": 1000
                    },
                    "requestId": {
                      "type": "string",
                      "maxLength": 80
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRule"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "TMS_SESSION"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN"
      }
    },
    "schemas": {
...
      "RuleVersion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "versionNo": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "contentJson": {
            "type": "string",
            "maxLength": 4000
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "publishedBy": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "nullable": true
          }
        },
        "required": []
      },
      "Ruleset": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "code": {
            "type": "string",
            "maxLength": 32
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "active": {
            "type": "boolean"
          },
          "activeVersion": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RuleVersion"
              }
            ]
          }
        },
        "required": []
      },
      "CreateRuleset": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INTERNAL_DEMO"
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "CreateRuleVersion": {
        "type": "object",
        "properties": {
          "contentJson": {
            "type": "string",
            "maxLength": 4000,
            "description": "JSON theo InternalRulePolicy; schemaVersion=1, scope=INTERNAL_DEMO; sourceReference, titlePrefix và đủ bảy requiredFields theo ADR-006."
          }
        },
        "required": [
          "contentJson"
        ]
      },
      "PublishRule": {
        "type": "object",
        "properties": {
          "expectedActiveVersionId": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "required": [
          "expectedActiveVersionId"
        ]
      },
...
    }
  }
}